update: separate json display from made
