+++ other/materialsproject/api_interoperability_showcase.py
@@ -6,23 +6,23 @@
 # </a>
 
 # # Overview
-# 
+#
 # This example was created as part of our [Advanced Topics Webinar](https://www.youtube.com/watch?v=psSFC409jSg) on February 19, 2021. This webinar focused on explaining our API in detail, and provided examples of many areas of its functionality.
-# 
+#
 # In this notebook, we showcase a major advantage of APIs: interoperability. We begin by performing a query using the [Materials Project](https://materialsproject.org) API for all systems containing Iron and Oxygen. We then filter our results (for demonstraiton purposes, we keep only the first 10 materials found). Finally, we upload our results to the Mat3ra platform, where further calculations could be performed to characterize these materials.
 
 # # Complete Authorization Form and Initialize Settings
-# 
+#
 # This will also determine environment and set all environment variables. We determine if we are using Jupyter Notebooks or Google Colab to run this tutorial.
-# 
+#
 # If you are running this notebook from Google Colab, Colab takes ~1 min to execute the following cell.
-# 
+#
 # ACCOUNT_ID and AUTH_TOKEN - Authentication parameters needed for when making requests to [Mat3ra.com's API Endpoints](https://docs.mat3ra.com/rest-api/endpoints/).
-# 
+#
 # MATERIALS_PROJECT_API_KEY - Authentication parameter needed for when making requests to [Material Project's API](https://materialsproject.org/open)
-# 
+#
 # ORGANIZATION_ID - Authentication parameter needed for when working with collaborative accounts https://docs.mat3ra.com/collaboration/organizations/overview/
-# 
+#
 # > <span style="color: orange">**NOTE**</span>: If you are running this notebook from Jupyter, the variables ACCOUNT_ID, AUTH_TOKEN, MATERIALS_PROJECT_API_KEY, and ORGANIZATION_ID should be set in the file [settings.json](../../utils/settings.json) if you need to use these variables. To obtain API token parameters, please see the following link to the documentation explaining how to get them: https://docs.mat3ra.com/accounts/ui/preferences/api/
 
 # In[ ]:
@@ -46,7 +46,9 @@ if "COLAB_JUPYTER_IP" in os.environ:
         )
     )
 
-    get_ipython().system('GIT_BRANCH="dev"; export GIT_BRANCH; curl -s "https://raw.githubusercontent.com/Exabyte-io/api-examples/${GIT_BRANCH}/scripts/env.sh" | bash')
+    get_ipython().system(
+        'GIT_BRANCH="dev"; export GIT_BRANCH; curl -s "https://raw.githubusercontent.com/Exabyte-io/api-examples/${GIT_BRANCH}/scripts/env.sh" | bash'
+    )
 
 
 # # Imports
@@ -55,16 +57,15 @@ if "COLAB_JUPYTER_IP" in os.environ:
 
 
 from utils.settings import ENDPOINT_ARGS, MATERIALS_PROJECT_API_KEY
-from utils.visualize import display_JSON
+from utils.io import display_JSON
 
 import ase.io
 from pymatgen.ext.matproj import MPRester
 
 from mat3ra.api_client.endpoints.materials import MaterialEndpoints
 
-
 # # Query the Materials Project
-# 
+#
 # We begin by using the Materials Project API implemented in [PyMatGen](https://pymatgen.org/pymatgen.ext.matproj.html) to perform a query for all systems containing Iron and Oxygen.
 
 # # Query Materials Project for all systems containing Iron and Oxygen
@@ -79,9 +80,9 @@ print(iron_oxides_ids)
 
 
 # # Filtering the Results
-# 
+#
 # This returns a lot of materials - 160 to be exact! In many cases, it is useful to filter down the number of materials. For example, we may want to exclude large unit cells that may be computationally intensive to study. Or we may want to restrict our results to only thermodynamically-stable oxides, by use of the material's [energy above hull](https://wiki.materialsproject.org/Glossary_of_Terms).
-# 
+#
 # As a basic example, here we only keep the first 10 iron oxides that the Materials Project API returned to us, and discard the other 150.
 
 # In[ ]:
@@ -93,7 +94,7 @@ print(some_iron_oxides)
 
 
 # # Bringing Materials Into the User Account
-# 
+#
 # Now that we have filtered the results from Materials Project down to just 10 structures, we may want to study them further with the computational models provided by Mat3ra. For example, we may be interested in leveraging a DFT code to find the structure with the largest band-gap, or perhaps we want to conduct a high-throughput screening of each material's surface energies.
 
 # In[ ]:
@@ -113,4 +114,3 @@ materials = exabyte_materials_api.import_from_materialsproject(MATERIALS_PROJECT
 materials_set = exabyte_materials_api.create_set({"name": "Some Iron Oxides"})
 for material in materials:
     exabyte_materials_api.move_to_set(material["_id"], "", materials_set["_id"])
-
