+++ other/materialsproject/import_materials_from_materialsproject.py
@@ -6,21 +6,21 @@
 # </a>
 
 # # Overview
-# 
+#
 # This example demonstrates how to import materials from the materials project database via [Material](https://docs.mat3ra.com/api/Material/post_materials_import) endpoint.
 
 # # Complete Authorization Form and Initialize Settings
-# 
+#
 # This will also determine environment and set all environment variables. We determine if we are using Jupyter Notebooks or Google Colab to run this tutorial.
-# 
+#
 # If you are running this notebook from Google Colab, Colab takes ~1 min to execute the following cell.
-# 
+#
 # ACCOUNT_ID and AUTH_TOKEN - Authentication parameters needed for when making requests to [Mat3ra.com's API Endpoints](https://docs.mat3ra.com/rest-api/endpoints/).
-# 
+#
 # MATERIALS_PROJECT_API_KEY - Authentication parameter needed for when making requests to [Material Project's API](https://materialsproject.org/open)
-# 
+#
 # ORGANIZATION_ID - Authentication parameter needed for when working with collaborative accounts https://docs.mat3ra.com/collaboration/organizations/overview/
-# 
+#
 # > <span style="color: orange">**NOTE**</span>: If you are running this notebook from Jupyter, the variables ACCOUNT_ID, AUTH_TOKEN, MATERIALS_PROJECT_API_KEY, and ORGANIZATION_ID should be set in the file [settings.json](../../utils/settings.json) if you need to use these variables. To obtain API token parameters, please see the following link to the documentation explaining how to get them: https://docs.mat3ra.com/accounts/ui/preferences/api/
 
 # In[ ]:
@@ -44,7 +44,9 @@ if "COLAB_JUPYTER_IP" in os.environ:
         )
     )
 
-    get_ipython().system('GIT_BRANCH="dev"; export GIT_BRANCH; curl -s "https://raw.githubusercontent.com/Exabyte-io/api-examples/${GIT_BRANCH}/scripts/env.sh" | bash')
+    get_ipython().system(
+        'GIT_BRANCH="dev"; export GIT_BRANCH; curl -s "https://raw.githubusercontent.com/Exabyte-io/api-examples/${GIT_BRANCH}/scripts/env.sh" | bash'
+    )
 
 
 # # Imports
@@ -53,15 +55,14 @@ if "COLAB_JUPYTER_IP" in os.environ:
 
 
 from utils.settings import ENDPOINT_ARGS, MATERIALS_PROJECT_API_KEY
-from utils.visualize import display_JSON
+from utils.io import display_JSON
 
 from mat3ra.api_client.endpoints.materials import MaterialEndpoints
 
-
 # ## Set Parameters
-# 
+#
 # - **MATERIALS_PROJECT_IDS**: a list of material IDs to be imported
-# 
+#
 # - **TAGS**: a list of [tags](https://docs.mat3ra.com/entities-general/data/#tags) to assign to imported materials
 
 # In[ ]:
@@ -72,7 +73,7 @@ TAGS = ["tag1", "tag2"]
 
 
 # ## Import materials
-# 
+#
 # Initialize `MaterialEndpoints` class and call `import_from_materialsproject` function to import materials.
 
 # In[ ]:
@@ -83,11 +84,10 @@ materials = endpoint.import_from_materialsproject(MATERIALS_PROJECT_API_KEY, MAT
 
 
 # ## Print imported materials
-# 
+#
 # Print the list of imported materials in pretty JSON below.
 
 # In[ ]:
 
 
 display_JSON(materials)
-
