+++ utils/visualize.py
@@ -1,8 +1,6 @@
 import io
 import json
-import os
 import time
-import uuid
 from enum import Enum
 from typing import Dict, List, Optional, Tuple, Union
 
@@ -13,74 +11,14 @@ from IPython.display import HTML, Javascript, display
 from mat3ra.made.material import Material
 from mat3ra.made.tools.convert import to_ase
 from mat3ra.utils.array import convert_to_array_if_not
-from pandas import DataFrame
-from pandas.io.formats.style import Styler
 from pydantic import BaseModel
 
-from utils import settings
+from utils.io import display_JSON
 
 
 class ViewersEnum(str, Enum):
     wave = "wave"
     ase = "ase"
-
-
-def dataframe_to_html(df: DataFrame, text_align: str = "center") -> Styler:
-    """
-    Converts Pandas dataframe to HTML.
-    See https://pandas.pydata.org/pandas-docs/stable/style.html for more information about styling.
-
-    Args:
-        df (pd.DataFrame): Pandas dataframe.
-        text_align (str): text align. Defaults to center.
-    """
-    styles = [
-        dict(selector="th", props=[("text-align", text_align)]),
-        dict(selector="td", props=[("text-align", text_align)]),
-    ]
-    return df.style.set_table_styles(styles)
-
-
-def display_JSON(
-    obj: Union[dict, list], interactive_viewer: bool = settings.use_interactive_JSON_viewer, level: int = 2
-) -> None:
-    """
-    Displays JSON, either interactively or via a text dump to Stdout.
-
-    The interactive viewer is based on https://github.com/mljar/mercury/blob/main/mercury/widgets/json.py.
-
-    Args:
-        obj (dict): Object to display as nicely-formatted JSON
-        interactive_viewer (bool): Whether to use the interactive viewer or not
-        level (int): The level to which the JSON should be expanded by default
-    """
-    if interactive_viewer:
-        if isinstance(obj, (dict, list)):
-            json_str = json.dumps(obj)
-        else:
-            json_str = obj
-
-        id = str(uuid.uuid4())
-
-        web_dir = os.path.join(os.path.dirname(__file__), "web")
-
-        with open(os.path.join(web_dir, "renderjson.css")) as fp:
-            css = fp.read()
-
-        with open(os.path.join(web_dir, "renderjson.js")) as fp:
-            js = fp.read()
-
-        display(HTML(f'<style>{css}</style><div id="{id}"></div>'))
-        display(
-            HTML(
-                f"<script>{js} "
-                f"renderjson.set_show_to_level({str(level)}); "
-                f'renderjson.set_icons("▸","▾"); '
-                f'document.getElementById("{id}").appendChild(renderjson({json_str}))</script>'
-            )
-        )
-    else:
-        print(json.dumps(obj, indent=4))
 
 
 def get_material_image(material: Material, title: str, rotation="0x,0y,0z", repetitions=[1, 1, 1]):
